requirements.txt
# Core dependencies
anthropic>=0.18.0
temporalio>=1.5.0
pydantic>=2.0.0

# Document processing
pypdf2>=3.0.0
pycryptodome>=3.20.0
pytesseract>=0.3.10
python-magic>=0.4.27
markdown>=3.5.0
pillow>=10.0.0

# Embeddings and ML
sentence-transformers>=2.5.0
numpy>=1.24.0
scikit-learn>=1.4.0

# CLI
click>=8.1.0
rich>=13.0.0
prompt-toolkit>=3.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
